requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.1
h11==0.14.0
idna==3.7
passlib==1.7.4
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
